paper-c3rl/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-c3rl"
version = "0.1.0"
description = "C3-RL paper workspace"
requires-python = ">=3.10"
dependencies = [
    "agent-rl-core",
    "pyyaml>=6.0",
    "numpy>=1.26",
    "torch>=2.4",
    "transformers>=4.47",
    "peft>=0.13",
]

[tool.setuptools.packages.find]
where = ["methods"]
